Refactor Block Cache into configuration and state modules
